Separate tests into individual files
makes it easier to maintain and understand.
